+++ apps/web/next.config.js
@@ -7,4 +7,7 @@ const withTM = require('next-transpile-modules')([
 
 module.exports = withTM({
   reactStrictMode: true,
+  images: {
+    domains: ['cdn.dribbble.com'],
+  },
 })