+++ demo/src/Row.js
@@ -1,6 +1,0 @@
-import styled from '@emotion/styled'
-
-export default styled.div`
-  display: flex;
-  flex-wrap: wrap;
-`