+++ apps/web/CHANGELOG.md
@@ -1,5 +1,11 @@
 # web
 
+## 0.0.2
+
+### Patch Changes
+
+- Fix build step including /dist folder
+
 ## 0.0.1
 
 ### Patch Changes