+++ packages/mask/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @reactour/mask
 
+## 1.0.2
+
+### Patch Changes
+
+- Fix build step including /dist folder
+- Updated dependencies
+  - @reactour/utils@0.4.1
+
 ## 1.0.1
 
 ### Patch Changes