+++ packages/tour/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @reactour/tour
 
+## 3.0.3
+
+### Patch Changes
+
+- Fix build step including /dist folder
+- Updated dependencies
+  - @reactour/mask@1.0.2
+  - @reactour/popover@1.0.2
+  - @reactour/utils@0.4.1
+
 ## 3.0.2
 
 ### Patch Changes